Move the api controller to its own package.
